# BACKEND/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atlas API (español)",
    "version": "1.0.0"
  },
  "servers": [{ "url": "http://localhost:4450" }],
  "paths": {
...
    "/usuarios": {
      "get": {
        "summary": "Listar usuarios",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Crear usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioCrear" }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Obtener usuario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "put": {
        "summary": "Actualizar usuario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UsuarioCrear" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Eliminar usuario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/empresas": {
      "get": {
        "summary": "Listar empresas",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Crear empresa",
        "description": "Acepta dos formatos: formato simple (solo empresa) o formato completo (empresa con usuario). Si se envía usuario y empresa, crea automáticamente el usuario, la empresa y la relación empresa-usuario con rol administrador.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  { "$ref": "#/components/schemas/Empresa" },
                  { "$ref": "#/components/schemas/EmpresaConUsuario" }
                ]
              }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/empresas/{id}": {
      "get": {
        "summary": "Obtener empresa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "put": {
        "summary": "Actualizar empresa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Empresa" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Eliminar empresa",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/roles": {
      "get": {
        "summary": "Listar roles",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Crear rol",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Rol" }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/roles/{id}": {
      "get": {
        "summary": "Obtener rol",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "put": {
        "summary": "Actualizar rol",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Rol" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Eliminar rol",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/empresa-usuarios": {
      "get": {
        "summary": "Listar vínculos empresa-usuarios",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Crear vínculo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmpresaUsuario" }
            }
          }
        },
        "responses": { "201": { "description": "Creado" } }
      }
    },
    "/empresa-usuarios/{id}": {
      "get": {
        "summary": "Obtener vínculo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "put": {
        "summary": "Actualizar vínculo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EmpresaUsuario" }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Eliminar vínculo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UsuarioCrear": {
        "type": "object",
        "properties": {
          "nombre_usuario": { "type": "string" },
          "contrasena_hash": { "type": "string" },
          "correo": { "type": "string" }
        },
        "required": ["nombre_usuario", "contrasena_hash", "correo"]
      },
...
      "Empresa": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string" },
          "direccion": { "type": "string" },
          "telefono": { "type": "string" },
          "correo": { "type": "string" }
        },
        "required": ["nombre", "correo"]
      },
      "Rol": {
        "type": "object",
        "properties": {
          "nombre": { "type": "string" }
        },
        "required": ["nombre"]
      },
      "EmpresaUsuario": {
        "type": "object",
        "properties": {
          "empresa_id": { "type": "integer" },
          "usuario_id": { "type": "integer" },
          "rol_id": { "type": "integer" }
        },
        "required": ["empresa_id", "usuario_id", "rol_id"]
      },
...
      "EmpresaConUsuario": {
        "type": "object",
        "properties": {
          "usuario": {
            "type": "object",
            "properties": {
              "nombre_usuario": { "type": "string" },
              "contrasena_hash": { "type": "string" },
              "correo": { "type": "string" }
            },
            "required": ["nombre_usuario", "contrasena_hash", "correo"]
          },
          "empresa": {
            "$ref": "#/components/schemas/Empresa"
          }
        },
        "required": ["usuario", "empresa"]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido del endpoint /autenticacion/login"
      }
    }
  }
}